select: Add select override that supports regular files.
As per POSIX, regular files should be always ready for I/O.
The idea is the same as for poll in 103d0c1f0080cac15e989910d021b038badb6fa1.

Note that after this commit the above poll change will go as it should
be not necessary (select now does all the job).