Switch to not returning success status from core init functions.
A success status is useless at the point where core initialization takes place
(e.g. upon DLL initialization) as there is no way for the application to
receive the status code from such places and/or recovering from a failure. We
prefer to use assert() instead as this will crash the application and provide
a meaningful error message on the console.