mmap: Significantly reduce memory footprint when mapping files in chunks.
Mapping files in chunks (small windows of fixed size) is used by e.g.
git to limit memory usage to a certain amount to be able to work with
very big files in low memory conditions (i.e. when these files would
not fit if mapped as a whole). The previous implementation caused
*more* memory used by this usage pattern which broke checkout of
large repositories with "Out of memory" error. See #29 for more info.