Significantly Improve speed of touch_pages().
This applies to the most frequent case when a single memory object is passed in
which is already committed. In such a case only one DosQueryMem call will be
done followed by an immediate exit. This saves a lot of milliseconds in case of
big memory objects (hundreds of megabytes).