Log trace messages to same file where assert goes.
It's still necessary to define TRACE_ENABLED to have trace statements do
something (which is the case in debug builds). Note that trace output now
goes to a file (in /var/log/libcx) by default. In order to redirect them
to the console LIBCX_TRACE_TO_CONSOLE must be set in the environment. Note
that doing so will direct assert messages to the console as well (so they
will appear twice unless LIBC_STRICT_DISABLED is set; this is a bugofeature
of kLIBC) and no log files will be created at all.