Implement tracking LIBC file descriptors.
This is a big change that reformats internal structures used to
track open files for LIBCx needs. In particular, LIBCx file
descriptions are now associated with LIBC file descriptors so that
the latter is closed, usage count of the former is decreased and
when there are no more descriptors using it (and no other uses
through e.g. mmap), the LIBCx file description is deleted. This
greatly reduces system resource usage. See #43 for more detail.