spawn2: Fix unexpected override of passed envp vars with current ones.
Also fixes duplicates in the resulting environment of the child process. Both problems would only take place if the child process was a kLIBC process (as in this case spawn seems to pass envp through shared memory rather than rely on DosExecPgm where the second occurrence would be ignored and silently removed).