tst-mmap: Add mmapping the same region twice.
This should succeed according to POSIX. On OS/2 this results in the same
address returned (reference counting is kept for it in order to properly
free memory when it's no longer needed). Some programs may freak out
(e.g. Qt's QFile::map implementation for Unix) but there is nothing wrong
with returning the same address from the POSIX POV so these programs should
be fixed.