spawn2: Return PID of wrapped child, not wrapper in P_2_THREADSAFE mode.
This is important for consistency and there are many situations when it's vital for the caller (e.g. sharing system resources with a given child PID).

Note that for this to work reliably it was also necessaryto override all the waitpid family of APIs (as well as DosWaitChild, for consistency). These
APIs can only wait on direct children. So supplying wrapped child PID would break them all. Now these APIs replace the wrapped child PID returned
by spawn2 with the wrapper PID (which is the direct child) on input and do the reverse on output.

This is also needed for #63.