spawn2: Accept ERROR_SMG_START_IN_BACKGROUND from DosStartSession.
It's a valid result, it only means that the process has been started in the background although the foreground was requested (this is because DosStartSession is not allowed to chagne the foreground session, only the operator can do that).

In particular, this fixes a case when tst_QProcess::startDetached and tst_QProcess::detachedProcessParameters in Qt5 are run in the background.