spawn2: Close stdio handles in P_2_THREADSAFE wrapper after starting child.
This is to give the child process full control over the open state of these handles. This is important if the parent wants to do something with a file bound to a standard handle (e.g. delete it) immediately after the target process closes the file. Having these handles open in the wrapper could make the parent fail because the wrapper would be still holding these handles open.

Needed for tst_QProcess::detachedProcessParameters.