Use realloc version that zeroes new memory.
This can fix a bunch of observed (and potential) crashes due to using arbitrary memory and such.

One of such crashes is documented in #46 (using uninitialized memory and/or freed memory blocks) and this commit provides a fix for it. Also, this commit fixes a possible mmap issue due to dirty page map being filled with garbage.