shmem: Unbound memory views from handles.
Process-specific memory views are now bound directly to memory objects rather than to handles (which could lead to separate view lists for the same object in the same process if there were duplicate handles). This allows to easily track when the last view is unmapped in the process and call DosFreeMem on it to instantly dereference the shared memory object instead of holding this reference till the process end w/o actually using the object. Another benefit is that it will allow to switch the memory object between read-write and read-only depending on whether there are read-write views or not (will be implemented in the next commit).

Needed for #83.