Implement interrupt_request service.
This service allows one LIBCx process to execute arbitrary LIBCx code in another LIBCx process with almost no delay (in both sync and async modes). This is a powerful feature that allows to do many fancy things in terms of process coordination. The service is implemented using the XCPT_SIGNAL_KILLPROC exception.

Needed for #87.