Fix assertions during LIBCx uninit.
This is a regression of 25dbd0193e555683d437e2a25849079ab5ca2b54 which could be seen in tst-mmap3, tst-mmap4 and other tests because ProcDesc was already gone by the time when it was still needed.

This commit also adds an optimization to get_proc_desc_ex that simply returns gpProcDesc for the current process when possible instead of querying the hash map.