Fix endless hang at startup when no UNIXROOT env. var is set.
The hang was caused by DLL uninitialization after a crash causing an unexpected recursion into the log creation routine when it was attempting to create a log file with no UNIXROOT variable present in the environment. The crash was caused by modifying a const variable located in a r/o segment.

UNIXROOT was missing because of bitwiseworks/libc#100.