Write assertions to log file even if no log groups are enabled.
This was a regression of edd1896 and would affect debug LIBCx builds where assertions were written only to console but not to the .log file, if the respective log group was disabled with LIBCX_TRACE. Note that the extended header info (LIBCx DLL version and path) would not end up in the log too in debug builds for the same reason.

Required some work in LIBC, check bitwiseworks/libc#113.

Note that assertions in release builds would also not appear but because LIBC would crash earlier due to another problem, see #98 and bitwiseworks/libc#112. Before that faulty change, release builds were not affected because they don't use log groups at all and hence messages couldn't be disabled with LIBCX_TRACE.

Closes #97.